Polished layout, created detail_extra_section_no_center_vertical.xml
For devices < sw360dp, the extras section is contained within a
ScrollView and it leaves white space at the bottom if we don't fill out
the screen. Therefore, we need to add padding to the bottom of the extras
section so it fills the screen leaving no white space for better
aesthetics. This is possible with "center vertical" version of xml, but
padding on top and bottom must be always equal or we risk obscuring some
data! Hence we have created the "no center vertical" version, so we can
have different sizing for our top and bottom padding.

Also added full mock data for better visualization of layouts.